third_party/WebKit/LayoutTests/svg/animations/animate-to-low-prio-frozen-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
PASS rectWidth is > 100
PASS 200 is >= rectWidth
PASS successfullyParsed is true

TEST COMPLETE
